// src/shared/lib/types/htmlPageInfo.ts
/**
 * Описание наличия предыдущей и следующей страниц для пагинации.
 */
export interface HasPrevNextPage {
  /** Есть ли следующая страница */
  next: boolean;
  /** Есть ли предыдущая страница */
  prev: boolean;
  /** Номер текущей страницы */
  current: number;
}